+++ src/app/tags/page.tsx
@@ -1,7 +1,8 @@
 import { Metadata } from "next";
 import ListHeading from "@/components/list-heading";
 import ListSection from "@/components/list-section";
-import TagList from "@/components/tag-list";
+import TagList from "@/components/tags/tag-list";
+import TagListItem from "@/components/tags/tag-list-item";
 import { getTags } from "@/service/tags";
 
 export default async function TagsPage() {
@@ -10,7 +11,13 @@ export default async function TagsPage() {
   return (
     <ListSection>
       <ListHeading>태그 목록</ListHeading>
-      <TagList tags={tags} />
+      {tags.length > 0 ? (
+        <TagList>
+          {tags.map((tag) => (
+            <TagListItem key={tag} tag={tag} link={`/posts?tag=${tag}`} />
+          ))}
+        </TagList>
+      ) : null}
     </ListSection>
   );
 }
