+++ src/components/post-article/post-article-header.tsx
@@ -1,6 +1,7 @@
 import { type Post } from "@/schema/posts";
 import formatDate from "@/utils/format-date";
-import TagList from "../tag-list";
+import TagList from "../tags/tag-list";
+import TagListItem from "../tags/tag-list-item";
 
 type Props = {
   post: Post;
@@ -10,7 +11,13 @@ export default function PostArticleHeader({ post }: Props) {
   return (
     <header className="flex flex-col gap-y-4">
       <h1 className="text-3xl sm:text-4xl font-bold">{post.title}</h1>
-      {post.tags ? <TagList tags={post.tags} /> : null}
+      {post.tags && post.tags.length > 0 ? (
+        <TagList>
+          {post.tags.map((tag) => (
+            <TagListItem key={tag} tag={tag} link={`/posts?tag=${tag}`} />
+          ))}
+        </TagList>
+      ) : null}
       <time className="text-sm text-zinc-700 dark:text-zinc-300">
         {formatDate(post.createdAt)} 작성
       </time>
